Added a Skill Screen where the player can manage his skills. The player can level up skills by spending XP points he gained by playing. A maximum of 10 skill points can be invested in any skill until the Drive Core becomes fully loaded. This forces the player to strategically choose his upgrades in order to help him survive as long as possible. At any time, a skill point can be refunded for a fraction of its original cost (except for the first skill point of every skill, which has an XP cost equal to its refund value).

+++ Includes/SaveLoad.h
@@ -16,7 +16,7 @@ struct SaveGame
 	unsigned dematerializeLevel;
 	unsigned overdriveLevel;
 	unsigned forceFieldLevel;
-	unsigned expPointsAvailable;
+	unsigned long expPointsAvailable;
 };
 
 class SaveLoad
